Use float samples internally.

+++ audioreader.h
@@ -5,6 +5,6 @@
 
 #include <stdint.h>
 
-bool read_audio_file(const char *filename, std::vector<int16_t> *samples, int *sample_rate);
+bool read_audio_file(const char *filename, std::vector<float> *samples, int *sample_rate);
 
 #endif  // !defined(_AUDIOREADER_H)
